requirements.txt
amqp==2.6.1
asgiref==3.4.1
beautifulsoup4==4.9.3
billiard==3.6.4.0
celery==4.4.7
certifi==2021.5.30
cffi==1.14.6
chardet==4.0.0
charset-normalizer==2.0.3
click==7.1.2
Django==2.2.24
django-admin-interface==0.16.3
django-adminlte3==0.1.6
django-bootstrap4==3.0.1
django-cors-headers==3.7.0
django-jazzmin==2.4.7
djangorestframework==3.12.4
gmaily==0.0.3
httplib2==0.19.1
idna==3.2
imap-tools==0.43.0
IMAPClient==2.2.0
imbox==0.9.8
importlib-metadata==3.10.0
Jinja2==3.0.1
keyring==23.0.1
kombu==4.6.11
MarkupSafe==2.0.1
mock==4.0.3
oauth2client==4.1.3
pyasn1==0.4.8
pyasn1-modules==0.2.8
pybase64==1.1.4
pycparser==2.20
PyNaCl==1.4.0
pyparsing==2.4.7
python-dateutil==2.8.2
pytz==2021.1
pywin32-ctypes==0.2.0
requests==2.26.0
rsa==4.7.2
six==1.15.0
soupsieve==2.2.1
sqlparse==0.4.1
structlog==21.1.0
uritemplate==3.0.1
urllib3==1.26.6
vine==1.3.0
zipp==3.4.1
